Коммит с исправлениями проекта
