+++ src/test/java/TestsHomeFAQTextChrome/TestPositiveMakeOrderButton.java
@@ -3,6 +3,7 @@ package TestsHomeFAQTextChrome;
 import io.github.bonigarcia.wdm.WebDriverManager;
 import org.hamcrest.MatcherAssert;
 import org.junit.After;
+import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
@@ -16,7 +17,7 @@ import pageobject.SecondFormPage;
 
 import static org.hamcrest.core.Is.is;
 @RunWith(Parameterized.class)
-public class TestPositiveMakeOrderButtonFromAbove {
+public class TestPositiveMakeOrderButton {
 
     private WebDriver driver;
     private final String name;
@@ -29,7 +30,7 @@ public class TestPositiveMakeOrderButtonFromAbove {
     private final String date;
     private final String day;
 
-    public TestPositiveMakeOrderButtonFromAbove(String name, String surName, String adress, String stationMetro, String namberStation, String phone, String date, String day) {
+    public TestPositiveMakeOrderButton(String name, String surName, String adress, String stationMetro, String namberStation, String phone, String date, String day) {
         this.name = name;
         this.surName = surName;
         this.adress = adress;
@@ -38,6 +39,16 @@ public class TestPositiveMakeOrderButtonFromAbove {
         this.phone = phone;
         this.date = date;
         this.day = day;
+    }
+    @Before
+    public void updown(){
+        WebDriverManager.chromedriver().setup();
+        ChromeOptions options = new ChromeOptions();
+        options.addArguments("--no-sandbox", "--disable-dev-shm-usage");
+
+        driver = new ChromeDriver(options);
+        // переход на страницу тестового приложения
+        driver.get("https://qa-scooter.praktikum-services.ru/");
     }
     @Parameterized.Parameters
     public static Object[][] getTestData(){
@@ -50,13 +61,6 @@ public class TestPositiveMakeOrderButtonFromAbove {
     @Test
     public void makeOrderButtonFromAbove() {
 
-        WebDriverManager.chromedriver().setup();
-        ChromeOptions options = new ChromeOptions();
-        options.addArguments("--no-sandbox", "--disable-dev-shm-usage");
-
-        driver = new ChromeDriver(options);
-        // переход на страницу тестового приложения
-        driver.get("https://qa-scooter.praktikum-services.ru/");
             String result = "Заказ оформлен";
           HomePageSamocat homePageSamocat = new HomePageSamocat(driver);
             homePageSamocat.clickToOrderFromAbove();
@@ -73,6 +77,30 @@ public class TestPositiveMakeOrderButtonFromAbove {
             secondFormPage.clickButtonOrder();
             secondFormPage.clickButtonYes();
             String expected = secondFormPage.checkTextModalWindow();
+        MatcherAssert.assertThat(result, is(expected));
+
+
+    }
+
+    @Test
+    public void makeOrderButtonBelow() {
+
+        String result = "Заказ оформлен";
+        HomePageSamocat homePageSamocat = new HomePageSamocat(driver);
+        homePageSamocat.clickToOrderBelow();
+        FirstFormPage firstFormPage = new FirstFormPage(driver);
+        firstFormPage.addName(name);
+        firstFormPage.addSurname(surName);
+        firstFormPage.addAdress(adress);
+        firstFormPage.addMetro(stationMetro, namberStation);
+        firstFormPage.addPhone(phone);
+        firstFormPage.clickButtonNext();
+        SecondFormPage secondFormPage = new SecondFormPage(driver);
+        secondFormPage.addDate(date, day);
+        secondFormPage.rentalPeriod();
+        secondFormPage.clickButtonOrder();
+        secondFormPage.clickButtonYes();
+        String expected = secondFormPage.checkTextModalWindow();
         MatcherAssert.assertThat(result, is(expected));
