--- src/test/java/TestsHomeFAQTextChrome/TestSeventhQuestion.java
@@ -1,43 +1,0 @@
-package TestsHomeFAQTextChrome;
-
-import io.github.bonigarcia.wdm.WebDriverManager;
-import org.hamcrest.MatcherAssert;
-import org.junit.After;
-import org.junit.Test;
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.chrome.ChromeDriver;
-import org.openqa.selenium.chrome.ChromeOptions;
-import pageobject.HomePageSamocat;
-
-import static org.hamcrest.core.Is.is;
-
-public class TestSeventhQuestion {
-    private WebDriver driver;
-    @Test
-    public void checkText() {
-
-        WebDriverManager.chromedriver().setup();
-        ChromeOptions options = new ChromeOptions();
-        options.addArguments("--no-sandbox", "--disable-dev-shm-usage");
-
-        driver = new ChromeDriver(options);
-        // переход на страницу тестового приложения
-        driver.get("https://qa-scooter.praktikum-services.ru/");
-
-        HomePageSamocat homePageSamocat = new HomePageSamocat(driver);
-        //Ожидаемый результат
-        String expected = "Да, пока самокат не привезли. Штрафа не будет, объяснительной записки тоже не попросим. Все же свои.";
-        homePageSamocat.scroolForElement7();
-        homePageSamocat.clickForElement7();
-        //Фактический результат
-        String result = homePageSamocat.getTextForElement7();
-        //Сравнение ОР с ФР
-        MatcherAssert.assertThat(result, is(expected));
-
-    }
-    @After
-    public void teardown() {
-        // Закрой браузер
-        driver.quit();
-    }
-}
